submissions/Oxide/admin/requirements.txt
Django==5.2.4
Pillow==10.1.0
djangorestframework==3.14.0